farmacia_interface/requirements.txt
# Core dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Computer Vision & ML
opencv-python>=4.8.0
ultralytics>=8.0.0
torch>=1.8.0
torchvision>=0.9.0
scikit-image>=0.18.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Barcode processing
pyzbar>=0.1.9
zxingcpp>=2.0.0

# Data analysis
pandas>=2.0.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.17.0

# API and web
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.4.0

# Utilities
pathlib2>=2.3.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development (optional)
black>=23.0.0
flake8>=6.0.0
